extract into standalone utils
