+++ packages/components/src/utils/isSupportedBrowser.ts
@@ -1,0 +1,17 @@
+import supportedBrowsers from "./supportedBrowsers"
+
+export interface SupportedBrowserBannerProps {
+  userAgent?: string | undefined
+}
+
+const getUserAgent = (): string | undefined => {
+  return typeof window !== "undefined" ? window.navigator.userAgent : undefined
+}
+
+export const isSupportedBrowser = ({
+  userAgent,
+}: SupportedBrowserBannerProps): boolean => {
+  if (userAgent === undefined) return true
+  if (userAgent === "") return true
+  return supportedBrowsers.test(userAgent || getUserAgent() || "")
+}
